Program uses QApplication to allow using GUI elements (file browser)
In order to streamline selecting a file to use with the program, a file browser will appear to allow the user to select it via a file structure instead of type out or paste the entire path.

QCoreApplication does not allow this, we need to move to QApplication.

We will also add gui elements to the project file too.