Added openGameTemplatesFile() to read in user file
When the user is prompted to select a file to be parsed the openGameTemplates function will be called to take that path and open and read in the selected file.

It will return a list of game template objects if they can be retrieved from the file. (Will return nullptr if it can not.)