Add parseGameTemplatesFile() to extract/retrieve GameTemplates from user file
Once the user has given a file it will be opened and read in, then the information from that file will be passed to parseGameTemplatesFile() in order to parse the file and retrieve a list of all of the game templates inside and return them. It will return an empty list if there are none.