Add promptForTemplatesFile to retrieve user's file to be opened
In order to open and parse the game templates file, the user will be prompted to select the file they want to use via a gui file browser.

This path will be passed to openGameTemplates file to check if it can be read. If it can it will be passed in order to be parsed. If both are successful it will return a list of game template objects. Else it will return  a nullptr.