Read hash in original endianness for consistency with other tools
The files that are read in have a reverse endianness. When we are reading it, we account for this so that when we read integers they are the correct values (so that we know how much data to read in and can accurately parse the file)

However, not all of the information is read in as a standard number.

Hashes can be read in with the same endianness, however when reading the file in a hex editor, it will appear in the file's original endianness. So if we read it in reverse and display it, it will not be familiar to a user who is using a hex editor.

So to avoid this, any hashes will be read in their original ordering for when we display it's raw hexadecimal. However, we will still consider endianness when displaying an ascii representation to maintain consistency with other editors.