Handle templates followed by 12 byte data structure
While the DLC GameTemplates.wsd file is read in as one would expect, there are issues with reading in the GameTemplates.wsd found in the loosefiles_binpc.pack in full.

This is because some template files have 12 bytes of data before the next template begins.

Currently it is unknown why these 12 bytes exist, or how to tell when they are about to show up. So after reading in each template, we will peek the next 12 bytes to check if they have the 0x08 0x00 0x00 0x00 0x00 0x00 0x00 0x00 0x00 0x00 0x00  0x00 signature.

If it is detected, these 12 bytes will be skipped and the next template will be read in as usual.