Handle trailing .pack data when parsing templates
Currently, the parsing function expects that once the final template is read in that the file should be fully read through.

However, this is not the case. There is a singular additional byte at the end of the pack. This causes the function to expect another template where there is none (and not enough data for one)

So in order to catch this scenario, it will skip to the end of the file if there is not enough remaining data for a template or the 12 byte template trailing structure handled in the previous commit.