Added loop to accept user commands to initiate list functions
Now that several listing functions have been added, the next step is to allow the user to call them.

The user can now enter the command list, then enter which kind of list task they want done.

Currently the options are simply "templates" to list all the templates that were loaded or list help, which lists the command options.