Improved spacing for printed instructions
The program will print basic commands to the user to instruct them on how to use the program. In order to improve readability, the spacing per line has been shifted so that some of these will not print on the immediate line, but the line after so that it is better separated.