Add function to list templates that have a specific hash with a specific data pairing
It is already possible to search for a template that contains a specific hash. However, if one has a specific hash paired with a specific piece of data that they want to search for, it was not possible.

Now a user can enter a hash and data to search for templates with that exact pairing.