Add function to search for templates of a specific subtype via a keyword
Added a function that allows a user to to search for and list all templates of a specific user given subtype that contain a user given keyword in their name.

This should allow the user to search for a particular word or set of letters and help keep the results only displaying relevant templates.