Simplified main loop exit cases
In the main function is a loop that allows the user to keep entering commands until they type exit to allow the program to terminate.

Additionally, some of the commands that can be entered allow to try again or put in a new option if the result fails to find anything.

The user is allowed to exit out of these nested loops. However, there is a lot of confusing naming conventions regarding this.

The main loop has an exit clause, each category of command has it's own exit clause, and some commands of those categories have an exit clause.

The exit clause of the category currently does not do anything and has been removed.

Further, the names of the exit clause variables for the commands have been changed.

This is because some had the same name as the main loop exit clause variable.

So depending on the scope you may be using the command's exit variable, but if you did not define one, you may be unintentionally setting the variable for the entire loop.

This caused numerous small bugs that caused the program to terminate after specific commands.

The names have been changed to make this less likely and removed redundancy.