Added function to list templates that use each model hash value
Templates may have a hash value to represent which model to use (or multiple).

Multiple templates may use the same model.

It is already possible to search a specific hash and value combination. However, this function allows searching every single hash value found in the file and list all of the templates that use each hash.